% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_stringeval.R
\name{any.empty}
\alias{any.empty}
\title{Checks if an array, a dataframe or a matrix contains any empty cell.}
\usage{
\method{any}{empty}(data)
}
\arguments{
\item{data}{An \code{array}, a \code{dataframe} or a \code{matrix}.}
}
\value{
A \code{logical}, TRUE if the array contains at least 1 empty cell,
        FALSE if it does not.
}
\description{
Checks if an array, a dataframe or a matrix contains any empty cell.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
