% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_stringeval.R
\name{dt.develop}
\alias{dt.develop}
\title{Develops a data.table row if it contains information for more than 1 sample.}
\usage{
dt.develop(dt.row)
}
\arguments{
\item{dt.row}{A \code{data.table} of 1 single row.}
}
\value{
A developped \code{data.table} of n rows, with 1 row per sample.
}
\description{
Develops a data.table row if it contains information for more than 1 sample.
}
\author{
Yoann Pageaud.
}
