% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TCGA_biotab.R
\name{get.ls.TCGA.biotab}
\alias{get.ls.TCGA.biotab}
\title{Returns a list of biotabs, one biotab per TCGA project.}
\usage{
get.ls.TCGA.biotab(proj.ids, dl.dir = NULL, rm.data = FALSE)
}
\arguments{
\item{proj.ids}{A \code{character} vector containing TCGA projects names.}

\item{dl.dir}{A \code{character} specifying the path where downloaded
clinical raw data should be saved. If NULL, a folder
'GDCdata' is created in the working directory
(Default: dl.dir = NULL).}

\item{rm.data}{A \code{logical} specifying whether to remove the downloaded
data or not once the biotab retrieved.}
}
\value{
A \code{list} containing results for each TCGA projects. A result is
        a \code{list} itself which contains 6 elements:
        \itemize{
         \item{biotab: a \code{data.frame} containing the project clinical
         data.}
         \item{description.table: a \code{data.frame} containing short
         description of the biotab column names.}
         \item{CDE_ID.table: a \code{data.frame} containing the Clinical Data
         Element IDs for each column of the biotab.}
         \item{warnings.query: a \code{list} of the possible warnings related
         to GDCquery().}
         \item{warnings.DL: a \code{list} of the possible warnings related to
         GDCdownload().}
         \item{warnings.preps: a \code{list} of the possible warnings related
         to GDCprepare().}
        }
}
\description{
Returns a list of biotabs, one biotab per TCGA project.
}
\author{
Yoann Pageaud.
}
