% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TCGA_biotab.R
\name{get.TCGA.clinical}
\alias{get.TCGA.clinical}
\title{Collects patients clinical data from specific TCGA projects.}
\usage{
get.TCGA.clinical(biotabs.dir, proj.id, patient.barcode, as.dt = FALSE)
}
\arguments{
\item{biotabs.dir}{A \code{character} to specify the directory where all
biotabs have been downloaded using
get.ls.TCGA.biotab() or get.TCGA.biotab().}

\item{proj.id}{A \code{character} vector of supported TCGA projects
and/or ICGC cohorts.}

\item{patient.barcode}{A \code{character} vector of valid TCGA patient IDs.}

\item{as.dt}{A \code{logical} to specify whether the output table
should be a tibble (Default: as.dt = FALSE) or a
data.table (as.dt = TRUE).}
}
\value{
A \code{list} with selected projects containing TCGA patients
        clinical data as a \code{tibble} or a \code{data.table}.
}
\description{
Collects patients clinical data from specific TCGA projects.
}
\author{
Yoann Pageaud.
}
