% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TCGA_biotab.R
\name{ICGC.to.TCGA.project}
\alias{ICGC.to.TCGA.project}
\title{Convert TCGA project ID into ICGC cohort and vice versa.}
\usage{
ICGC.to.TCGA.project(project = NULL, to = "TCGA")
}
\arguments{
\item{project}{A \code{character} vector of valid TCGA projects and/or
ICGC cohorts.}

\item{to}{A \code{character} specifying the output format of project
names
(Default: to = "TCGA"; Supported: to = c("TCGA", "ICGC")).}
}
\value{
A \code{character} vector of the formated project names.
}
\description{
Convert TCGA project ID into ICGC cohort and vice versa.
}
\author{
Yoann Pageaud.
}
