% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_stringeval.R
\name{is.upper}
\alias{is.upper}
\title{Checks if an alphanumerical string contains any upper case character.}
\usage{
is.upper(str)
}
\arguments{
\item{str}{A \code{character} vector to be evaluated looking for any
uppercase character.}
}
\value{
A \code{logical}.
}
\description{
Checks if an alphanumerical string contains any upper case character.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
