% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_biotabs.R
\name{multi.subset.biotab}
\alias{multi.subset.biotab}
\title{Multi-step clinical biotabs subsetting following a type of data and its
manifest.}
\usage{
multi.subset.biotab(Biotab.RDS, list.manifests)
}
\arguments{
\item{Biotab.RDS}{A \code{list} of tibbles containing clinical data.}

\item{list.manifests}{A \code{dataframe} vector containing multiple manifests
in a given order to successively subset clinical data
following the different manifests.}
}
\value{
A \code{list} with biotabs as \code{tibbles} and a summary table as a
        \code{data.frame} of the different steps of subset.
}
\description{
Multi-step clinical biotabs subsetting following a type of data and its
manifest.
}
\author{
Yoann Pageaud.
}
