% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funRDS.R
\name{saveRDSls}
\alias{saveRDSls}
\title{Save a list of named object as independent RDS files.}
\usage{
saveRDSls(
  list.object,
  dir.save,
  ascii = FALSE,
  version = NULL,
  compress = TRUE,
  refhook = NULL
)
}
\arguments{
\item{list.object}{A \code{list} of named object to serialize. Object names
are used as file names.}

\item{dir.save}{A \code{character} to specify the path to a folder where
the RDS files should be saved.}

\item{ascii}{A \code{logical}. If TRUE or NA, an ASCII representation
is written; otherwise (default), a binary one is used.}

\item{version}{The workspace format version to use. NULL specifies the
current default version (3). The only other supported
value is 2, the default from R 1.4.0 to R 3.5.0.}

\item{compress}{A logical specifying whether saving to a named file is to
use "gzip" compression, or one of "gzip", "bzip2" or "xz"
to indicate the type of compression to be used. Ignored if
file is a connection.}

\item{refhook}{A hook function for handling reference objects.}
}
\value{
A \code{type} object returned description.
}
\description{
Save a list of named object as independent RDS files.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
