% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_biotabs.R
\name{subset.clinical.biotab}
\alias{subset.clinical.biotab}
\title{Subset clinical biotabs following a type of data and its manifest.}
\usage{
\method{subset}{clinical.biotab}(Biotab.RDS, Manifest, data.type)
}
\arguments{
\item{Biotab.RDS}{A \code{list} of tibbles containing clinical data.}

\item{Manifest}{A \code{data.frame} containing TCGA information about
samples and data to use for subsetting clinical data}

\item{data.type}{A \code{character} to specify the type of data to use.}
}
\value{
A \code{list} with biotabs as \code{tibbles} and a summary table as a
        \code{data.frame} of the subset.
}
\description{
Subset clinical biotabs following a type of data and its manifest.
}
\author{
Yoann Pageaud.
}
