% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TCGA_biotab.R
\name{TCGA.clinical.as.dt}
\alias{TCGA.clinical.as.dt}
\title{Subset clinical data from collected patients clinical data, and return it
into a data.table.}
\usage{
TCGA.clinical.as.dt(
  TCGA.clinical,
  biotab.type = "clinical_patient",
  columns = "all",
  to.ICGC = FALSE
)
}
\arguments{
\item{TCGA.clinical}{A \code{list} of TCGA projects clinical data generated
using get.TCGA.clinical().}

\item{biotab.type}{A \code{character} to specify the type of biotab to
extract data from
(Default: biotab.type = "clinical_patient"). Supported
biotab.type can match any table name without the project
suffix.}

\item{columns}{A \code{character} vector specifying the column names
you want to retrieve clinical data from. Columns name
can appear in some projects only. If columns = 'all',
all existing columns throughout projects listed will be
retrieved.}

\item{to.ICGC}{A \code{logical} to specify whether the TCGA project
should be converted into ICGC cohorts (to.ICGC = TRUE)
or not (Default: to.ICGC = FALSE).}
}
\value{
A \code{data.table} containing data from columns of interest, for the
        type of biotab specified and for the TCGA patients available in
        'TCGA.clinical'.
}
\description{
Subset clinical data from collected patients clinical data, and return it
into a data.table.
}
\author{
Yoann Pageaud.
}
